requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
pydantic>=2.0.0
pyyaml>=6.0
pytest>=7.0.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
